Cloud Function/requirements.txt
functions-framework==3.*
requests==2.31.0
jwt==1.3.1
google-auth
google-cloud-secret-manager
google-cloud-logging
google-cloud-storage==2.14.0